refactor: move chartconfig file
