+++ lib/configs/chartConfig.ts
