+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@clerk/nextjs": "^6.12.5",
+    "@google-cloud/storage": "^7.15.2",
     "@hookform/resolvers": "^4.1.3",
     "@prisma/client": "^6.5.0",
     "@radix-ui/react-alert-dialog": "^1.1.6",
@@ -32,6 +33,7 @@
     "framer-motion": "^12.5.0",
     "ioredis": "^5.6.0",
     "lucide-react": "^0.482.0",
+    "multer": "^1.4.5-lts.2",
     "next": "^15.2.4",
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
@@ -43,6 +45,7 @@
     "svix": "^1.61.4",
     "tailwind-merge": "^3.0.2",
     "tailwindcss-animate": "^1.0.7",
+    "uuid": "^11.1.0",
     "zod": "^3.24.2"
   },
   "devDependencies": {
